Grenade and Enemy additions/updates
Added a grenade warning icon and added UI code to trigger it in GameManager and Grenade.

Added a particle effect to the grenade to create an explosion.

Fixed an enemy ai bug where the enemy could still be shot while dying, causing the updateGameGoal function to continue being called and ending the game prematurely.